+++ src/main/java/me/despical/kotl/commands/SubCommand.java
@@ -20,7 +20,6 @@ package me.despical.kotl.commands;
 
 import me.despical.kotl.Main;
 import me.despical.kotl.commands.exception.CommandException;
-import me.despical.kotl.handlers.ChatManager;
 import org.bukkit.command.CommandSender;
 import org.bukkit.plugin.java.JavaPlugin;
 
@@ -64,7 +63,7 @@ public abstract class SubCommand {
 
 	public abstract int getMinimumArguments();
 
-	public abstract void execute(CommandSender sender, ChatManager chatManager, String[] args) throws CommandException;
+	public abstract void execute(CommandSender sender, String[] args) throws CommandException;
 	
 	public abstract List<String> getTutorial();