+++ src/main/java/me/despical/kotl/commands/admin/ListCommand.java
@@ -21,7 +21,6 @@ package me.despical.kotl.commands.admin;
 import me.despical.kotl.arena.Arena;
 import me.despical.kotl.arena.ArenaRegistry;
 import me.despical.kotl.commands.SubCommand;
-import me.despical.kotl.handlers.ChatManager;
 import org.bukkit.command.CommandSender;
 
 import java.util.Collections;
@@ -52,15 +51,14 @@ public class ListCommand extends SubCommand {
 	}
 
 	@Override
-	public void execute(CommandSender sender, ChatManager chatManager, String[] args) {
+	public void execute(CommandSender sender, String[] args) {
 		if (ArenaRegistry.getArenas().isEmpty()) {
-			sender.sendMessage(chatManager.getPrefix() + chatManager.colorMessage("Commands.List-Command.No-Arenas-Created"));
+			sender.sendMessage(plugin.getChatManager().getPrefix() + plugin.getChatManager().colorMessage("Commands.List-Command.No-Arenas-Created"));
 			return;
 		}
 
 		List<String> arenas = ArenaRegistry.getArenas().stream().map(Arena::getId).collect(Collectors.toList());
-		sender.sendMessage(chatManager.getPrefix() + chatManager.colorMessage("Commands.List-Command.Format").replace("%list%",
-			arenas.toString().substring(1, arenas.toString().length() - 1)));
+		sender.sendMessage(plugin.getChatManager().getPrefix() + plugin.getChatManager().colorMessage("Commands.Admin-Commands.List-Command.Format").replace("%list%", String.join(", ", arenas)));
 	}
 
 	@Override