+++ src/main/java/me/despical/kotl/commands/admin/arena/EditCommand.java
@@ -21,7 +21,6 @@ package me.despical.kotl.commands.admin.arena;
 import me.despical.kotl.arena.Arena;
 import me.despical.kotl.arena.ArenaRegistry;
 import me.despical.kotl.commands.SubCommand;
-import me.despical.kotl.handlers.ChatManager;
 import me.despical.kotl.handlers.setup.SetupInventory;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
@@ -53,12 +52,12 @@ public class EditCommand extends SubCommand {
 	}
 
 	@Override
-	public void execute(CommandSender sender, ChatManager chatManager, String[] args) {
+	public void execute(CommandSender sender, String[] args) {
 		Player player = (Player) sender;
 		Arena arena = ArenaRegistry.getArena(args[0]);
 		
 		if (arena == null) {
-			player.sendMessage(chatManager.getPrefix() + chatManager.colorMessage("Commands.No-Arena-Like-That"));
+			player.sendMessage(plugin.getChatManager().getPrefix() + plugin.getChatManager().colorMessage("Commands.No-Arena-Like-That"));
 			return;
 		}