+++ src/main/java/me/despical/kotl/commands/game/LeaderBoardCommand.java
@@ -21,7 +21,6 @@ package me.despical.kotl.commands.game;
 import me.despical.kotl.ConfigPreferences;
 import me.despical.kotl.api.StatsStorage;
 import me.despical.kotl.commands.SubCommand;
-import me.despical.kotl.handlers.ChatManager;
 import me.despical.kotl.user.data.MysqlManager;
 import org.apache.commons.lang.StringUtils;
 import org.bukkit.Bukkit;
@@ -57,9 +56,9 @@ public class LeaderBoardCommand extends SubCommand {
 	}
 
 	@Override
-	public void execute(CommandSender sender, ChatManager chatManager, String[] args) {
+	public void execute(CommandSender sender, String[] args) {
 		if (args.length == 0) {
-			sender.sendMessage(chatManager.getPrefix() + chatManager.colorMessage("Commands.Statistics.Type-Name"));
+			sender.sendMessage(plugin.getChatManager().getPrefix() + plugin.getChatManager().colorMessage("Commands.Statistics.Type-Name"));
 			return;
 		}
 
@@ -67,7 +66,7 @@ public class LeaderBoardCommand extends SubCommand {
 			StatsStorage.StatisticType statisticType = StatsStorage.StatisticType.valueOf(args[0].toUpperCase(java.util.Locale.ENGLISH));
 			printLeaderboard(sender, statisticType);
 		} catch (IllegalArgumentException exception) {
-			sender.sendMessage(chatManager.getPrefix() + chatManager.colorMessage("Commands.Statistics.Invalid-Name"));
+			sender.sendMessage(plugin.getChatManager().getPrefix() + plugin.getChatManager().colorMessage("Commands.Statistics.Invalid-Name"));
 		}
 	}