+++ src/main/java/me/despical/kotl/events/Events.java
@@ -71,7 +71,7 @@ public class Events implements Listener {
 			}
 		}
 
-		if (event.getPlayer().isOp() || event.getPlayer().hasPermission("kotl.admin") || event.getPlayer().hasPermission("kotl.command.override")) {
+		if (event.getPlayer().isOp() || event.getPlayer().hasPermission("kotl.admin")) {
 			return;
 		}